+++ KeeAnywhere/Configuration/ConfigurationService.cs
@@ -1,8 +1,10 @@
 using System;
 using System.Collections.Generic;
+using System.IO;
 using System.Linq;
 using System.Reflection;
 using KeeAnywhere.StorageProviders;
+using KeePass.App.Configuration;
 using KeePass.Plugins;
 using KeePassLib.Utility;
 using Newtonsoft.Json;
@@ -149,6 +151,12 @@ namespace KeeAnywhere.Configuration
             if (PluginConfiguration == null)
             {
                 this.PluginConfiguration = new PluginConfiguration();
+            }
+
+            if (string.IsNullOrEmpty(this.PluginConfiguration.BackupToLocalFolder))
+            {
+                this.PluginConfiguration.BackupToLocalFolder = Path.Combine(AppConfigSerializer.LocalAppDataDirectory,
+                    "KeeAnywhereBackups");
             }
         }