+++ KeeAnywhere/Configuration/PluginConfiguration.cs
@@ -1,5 +1,8 @@
 using System;
+using System.ComponentModel;
+using System.IO;
 using System.Runtime.Serialization;
+using KeePass.App.Configuration;
 using Newtonsoft.Json;
 using Newtonsoft.Json.Converters;
 
@@ -8,8 +11,48 @@ namespace KeeAnywhere.Configuration
     [DataContract]
     public class PluginConfiguration
     {
+        private string _backupToLocalFolder;
+
         [DataMember]
         public bool IsOfflineCacheEnabled { get; set; }
+
+        public string OfflineCacheFolder
+        {
+            get
+            {
+                return Path.Combine(AppConfigSerializer.LocalAppDataDirectory, "KeeAnywhereOfflineCache");
+            }
+        }
+
+        [DataMember]
+        public bool IsBackupToLocalEnabled { get; set; }
+
+        [DataMember]
+        public bool IsBackupToRemoteEnabled { get; set; }
+
+        [DataMember]
+        public string BackupToLocalFolder
+        {
+            get { return _backupToLocalFolder; }
+            set
+            {
+                _backupToLocalFolder = value;
+                if (this.IsBackupToLocalEnabled && !string.IsNullOrEmpty(_backupToLocalFolder) && !Directory.Exists(_backupToLocalFolder))
+                {
+                    try
+                    {
+                        Directory.CreateDirectory(_backupToLocalFolder);
+                    }
+                    catch (Exception)
+                    { }
+                }
+            }
+        }
+
+        [DataMember]
+        [DefaultValue(10)]
+        [JsonProperty(DefaultValueHandling = DefaultValueHandling.Populate)]
+        public int BackupCopies { get; set; }
 
         [DataMember]
         [JsonConverter(typeof(StringEnumConverter))]