+++ KeeAnywhere/StorageProviders/StorageService.cs
@@ -1,20 +1,28 @@
 using System;
 using System.Linq;
+using System.Net;
 using System.Threading.Tasks;
 using KeeAnywhere.Configuration;
+using KeeAnywhere.Offline;
 using KeeAnywhere.WebRequest;
+using KeePass.Forms;
+using KeePass.Plugins;
 using KeePassLib.Serialization;
 
 namespace KeeAnywhere.StorageProviders
 {
-    public class StorageService
+    public class StorageService: IWebRequestCreate
     {
         private readonly ConfigurationService _configService;
+        private readonly CacheManagerService _cacheManagerService;
 
-        public StorageService(ConfigurationService configService)
+        public StorageService(ConfigurationService configService, CacheManagerService cacheManagerService)
         {
             if (configService == null) throw new ArgumentNullException("configService");
+            if (cacheManagerService == null) throw new ArgumentNullException("cacheManagerService");
+
             _configService = configService;
+            _cacheManagerService = cacheManagerService;
         }
 
         public IStorageProvider GetProviderByUri(StorageUri uri)
@@ -67,14 +75,33 @@ namespace KeeAnywhere.StorageProviders
             return account;
         }
 
+        public System.Net.WebRequest Create(Uri uri)
+        {
+            var providerUri = new StorageUri(uri);
+            var provider = this.GetProviderByUri(providerUri);
 
-        public void Register()
+            if (_configService.PluginConfiguration.IsOfflineCacheEnabled)
+            {
+                provider = _cacheManagerService.GetCachedProvider(provider, uri);
+            }
+
+            var itemPath = providerUri.GetPath();
+
+            return new KeeAnywhereWebRequest(provider, itemPath);
+        }
+
+        public void RegisterPrefixes()
         {
             foreach (var descriptor in StorageRegistry.Descriptors)
             {
                 FileTransactionEx.Configure(descriptor.Scheme, false);
-                System.Net.WebRequest.RegisterPrefix(descriptor.Scheme + ":", new KeeAnywhereWebRequestCreator(this));
+                System.Net.WebRequest.RegisterPrefix(descriptor.Scheme + ":", this);
             }
         }
+
+        
+
+       
+
     }
 }