+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ### New
 
+- \#5 Cache databases for disconnected (offline) usage
+- \#61 Simple Automatic Backup (Remote and/or Local)
+
 ### Fixed
 
 - \#58 Amazon S3: Cannot save/write database file to root of bucket